refactor: Separate generic and KIT ilias functions
